Create initial module using module boilerplate
Create project from our module boilerplate template. Pproject has been
downgraded to target RC1 as our immediate need is to get it working for
that version of Orchard Core. Slight change to the directory structure
too, the module has been placed in to it's own directory as we'll have
another project within this repo for generating tiles.