Initial development for adding POIs to map
Still a work in progress.

Added "POIs" tab when editing map that displays a preview of the map
tiles that the user can click on to add POIs. Selecting the map will
display a dialog to the user where they can choose the POI. The
available POIs are determined by fetching all content types with
stereotype of "POI". When the user selects a POI it'll display the
marker determined by settings on the PoiPart.