Fix issue where tiles could become blurry
Workaround that was being used to remove visible lines between tiles
doesn't cater for the fact that deep zoom will generate tiles of
different sizes. This would cause edge tiles to become blurred and not
align with adjacent tiles.

Within the leaflet project there is an issue that exposes a solution
that resizes the tiles by referencing the existing size of the tile.

https://github.com/Leaflet/Leaflet/issues/3575#issuecomment-150544739