Add GA tracking for when POI is clicked
If GA is enabled on the page, if a user selects a POI a custom event
will be sent to GA. Content editors can manage the category/action of
the event via a new "Analytics" tab when creating/editing a map. GA
event will also contain the title of the POI as the label.