Add ability to change marker dimensions on zoom
Added zoom ratio setting to `PoiPart` that is used to determine the
dimensions of the marker icon. This calculation is done by using the
height/width then add the number of zoom levels multiplied by the ratio.
For example, if the minimum zoom is `10`, the width & height is `10`,
the ratio is `1.5`, if the zoom level is `12`, the width & height will
be `20`.