Add deep linking for zoom & poi selection
When zooming or selecting a POI the URL will be updated so when the
user refreshes or shares a link they will see that zoom level and/or the
select POI already displaying.

There is a known issue that when closing via the close icon on the POI
a hash link will be added, which messes with the history state and can
cause the back button to exhibit a weird behaviour. The next version of
Leaflet (1.9) will resolve this issue.